ABS-26: when user is deactivated, all user bookings are disabled
