+++ src/main/java/com/acs/bookingsystem/security/CustomAuthenticationEntryPoint.java
@@ -2,7 +2,6 @@ package com.acs.bookingsystem.security;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.core.AuthenticationException;
@@ -10,9 +9,11 @@ import org.springframework.web.servlet.HandlerExceptionResolver;
 
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
-    @Autowired
-    @Qualifier("handlerExceptionResolver")
-    private HandlerExceptionResolver resolver;
+    private final HandlerExceptionResolver resolver;
+
+    public CustomAuthenticationEntryPoint(@Qualifier("handlerExceptionResolver") HandlerExceptionResolver resolver) {
+        this.resolver = resolver;
+    }
 
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) {
