+++ src/test/java/com/acs/bookingsystem/danceclass/service/DanceClassServiceTest.java
@@ -2,7 +2,6 @@ package com.acs.bookingsystem.danceclass.service;
 
 import com.acs.bookingsystem.danceclass.DanceClassTestData;
 import com.acs.bookingsystem.danceclass.entity.DanceClass;
-import com.acs.bookingsystem.danceclass.mapper.DanceClassMapper;
 import com.acs.bookingsystem.danceclass.repository.DanceClassRepository;
 import com.acs.bookingsystem.danceclass.request.DanceClassRequest;
 import com.acs.bookingsystem.danceclass.service.impl.DanceClassService;
@@ -30,9 +29,6 @@ class DanceClassServiceTest {
     @Mock
     private DanceClassRepository danceClassRepository;
 
-    @Mock
-    private DanceClassMapper danceClassMapper;
-
     @InjectMocks
     private DanceClassService danceClassService;
 
@@ -49,7 +45,7 @@ class DanceClassServiceTest {
     }
 
     @Test
-    void whenUserAdmin_createDanceClass_danceClassDTOReturned() {
+    void whenUserAdmin_createDanceClass_danceClassReturned() {
         when(danceClassRepository.findByActiveIsTrueAndClassType(any())).thenReturn(Optional.ofNullable(danceClass));
         when(danceClassRepository.save(any())).thenReturn(danceClass);
