--- src/test/java/com/acs/bookingsystem/user/service/impl/UserServiceImplTest.java
@@ -1,28 +1,0 @@
-package com.acs.bookingsystem.user.service.impl;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class UserServiceImplTest {
-
-    @Test
-    void updateUserCredentials() {
-    }
-
-    @Test
-    void getUserProfile() {
-    }
-
-    @Test
-    void getUserProfiles() {
-    }
-
-    @Test
-    void updateUserStatus() {
-    }
-
-    @Test
-    void resetPassword() {
-    }
-}
